Rename TradeMenuUpdater
