+++ src/cow-react/modules/trade/index.ts
@@ -1,1 +1,1 @@
-export * from './updaters/LimitOrdersMenuUpdater'
+export * from './updaters/TradeMenuUpdater'
