+++ src/cow-react/modules/trade/updaters/TradeMenuUpdater/index.tsx
